# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Bibliology API
  description: |
    Bilingual (EN/FR) Bible study learning management system API.

    ## Authentication
    All protected endpoints require a JWT token in an HttpOnly cookie named `accessToken`.

    ## Rate Limiting
    - Anonymous: 10 requests/minute
    - Authenticated: 100 requests/minute

    ## Pagination
    List endpoints support pagination with `page` and `limit` query parameters.

  version: 1.0.0
  contact:
    name: Bibliology Support
    email: support@bibliology.com
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api-staging.bibliology.com
    description: Staging environment
  - url: https://api.bibliology.com
    description: Production environment
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, role]
              properties:
                email:
                  type: string
                  format: email
                  example: student@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePassword123!
                name:
                  type: string
                  example: John Doe
                role:
                  type: string
                  enum: [STUDENT, TEACHER]
                  example: STUDENT
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: accessToken=eyJhbGc...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: accessToken=; HttpOnly; Secure; SameSite=Strict; Max-Age=0
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  # ============================================================================
  # COURSES
  # ============================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT token stored in HttpOnly cookie
...
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [STUDENT, TEACHER, ADMIN]
        avatar_url:
          type: string
        language_pref:
          type: string
          enum: [en, fr]
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - cookieAuth: []